Use BIGINT in DB for adjudicationId type

+++ src/main/resources/migration/common/V7__adjudication_id.sql
@@ -1,1 +1,1 @@
-ALTER TABLE additional_days_awarded ALTER COLUMN adjudication_id TYPE INTEGER USING adjudication_id::integer;
+ALTER TABLE additional_days_awarded ALTER COLUMN adjudication_id TYPE BIGINT USING adjudication_id::BIGINT;
